Port recent changes in MSVC build to Borland C build. Fix minor compilation problems in the components that has not been compiled by Borland C before.
DETAILS:
Although nobody seems to use Borland C, having an alternative compiler
on Windows helps to separatr MSVC errors from Microsoft SDK quirks.

DIFFERENCES:
None.

git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@8037 a1074d23-0009-0410-80fe-cf8c14f379e6

+++ gs/src/dwuninst.cpp
@@ -833,7 +833,7 @@ init(void)
 	    s++;
     }
     len = s - p;
-    strncpy(szLogFile, p, min(len, sizeof(szLogFile)-1));
+    strncpy(szLogFile, p, min(len, (int)sizeof(szLogFile)-1));
     szLogFile[len] = '\0';
     if (inquote && (*s == '\042')) {
 	s++;
